Values in KisPaintInformation are interpolated between event's paint info for each dab, time should be decimal rather than integer to be correctly interpolated on high velocity strokes.
Note : this patch is not enough to completely fix the speed sensor as
       KisPaintInformation's timestamps are still garbage.

+++ krita/image/brushengine/kis_paint_information.h
@@ -85,7 +85,7 @@ public:
                         qreal rotation = 0.0,
                         qreal tangentialPressure = 0.0,
                         qreal perspective = 1.0,
-                        int   time = 0);
+                        qreal time = 0.0);
 
     KisPaintInformation(const KisPaintInformation& rhs);
 
@@ -168,7 +168,7 @@ public:
     qreal perspective() const;
 
     /// Number of ms since the beginning of the stroke
-    int currentTime() const;
+    qreal currentTime() const;
 
     /**
      * The paint information may be generated not only during real
